making our test app more friendly to test different markup

+++ Program.cs
@@ -1,5 +1,6 @@
 ﻿using StaticWebEpiserverPlugin.RequiredCssOnly.Services;
 using System.Collections.Generic;
+using System.Text.RegularExpressions;
 
 namespace RequiredCssOnlyApp
 {
@@ -11,34 +12,45 @@ namespace RequiredCssOnlyApp
 
             RequiredCssOnlyService service = new RequiredCssOnlyService();
 
-            var html = System.IO.File.ReadAllText(folder + "page.html", System.Text.Encoding.UTF8);
+            var pageName = "page.html";
+            var html = System.IO.File.ReadAllText(folder + pageName, System.Text.Encoding.UTF8);
 
-            var cssFiles = new List<string>(new []{
-                @"bootstrap.css",
-                @"bootstrap-responsive.css",
-                @"style.css",
-                @"editmode.css",
-                @"media.css"
-            });
-
-            foreach (string fileName in cssFiles)
+            var linkMatches = Regex.Matches(html, "(?<link><link.*href=\"(?<url>[^\"]+)\"[^>]*>)");
+            foreach (Match linkMatch in linkMatches)
             {
+                var linkGroup = linkMatch.Groups["link"];
+                if (!linkGroup.Success)
+                {
+                    continue;
+                }
+
+                var urlGroup = linkMatch.Groups["url"];
+                if (!urlGroup.Success)
+                {
+                    continue;
+                }
+
+                if (string.IsNullOrEmpty(urlGroup.Value) || !urlGroup.Value.EndsWith(".css"))
+                {
+                    continue;
+                }
+
+                var fileName = urlGroup.Value;
+
                 var cssContent = System.IO.File.ReadAllText(folder + fileName, System.Text.Encoding.UTF8);
                 var resultingCss = service.RemoveUnusedRules(cssContent, html);
 
-                System.IO.File.WriteAllText(folder + fileName + "-out.css", resultingCss, System.Text.Encoding.UTF8);
+                System.IO.File.WriteAllText(folder + fileName.Replace(".css", "-out.css"), resultingCss, System.Text.Encoding.UTF8);
 
                 if (!string.IsNullOrEmpty(resultingCss))
                 {
-                    html = html.Replace($"<link href=\"" + fileName + "\" rel=\"stylesheet\" />", $"<style>{resultingCss}</style>");
+                    resultingCss = $"<style>{resultingCss}</style>";
                 }
-                else
-                {
-                    html = html.Replace($"<link href=\"" + fileName + "\" rel=\"stylesheet\" />", string.Empty);
-                }
+
+                html = html.Replace(linkGroup.Value, resultingCss);
             }
 
-            System.IO.File.WriteAllText(folder + @"page.html-out.html", html, System.Text.Encoding.UTF8);
+            System.IO.File.WriteAllText(folder + pageName.Replace(".html", "-out.html"), html, System.Text.Encoding.UTF8);
         }
     }
 }
